Reorganize lib
