+++ dggrs/src/bindings/mod.rs
@@ -7,4 +7,4 @@
 // discretion. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-pub mod wasm;
+pub mod napi;
