+++ dggrs/src/bindings/napi/package.json
@@ -1,0 +1,5 @@
+{
+  "devDependencies": {
+    "@napi-rs/cli": "^3.1.5"
+  }
+}
