--- dggrs/src/bindings/wasm/adapters/mod.rs
@@ -1,10 +1,0 @@
-// Copyright 2025 contributors to the GeoPlegma project.
-// Originally authored by Michael Jendryke (GeoInsight GmbH, michael.jendryke@geoinsight.ai)
-//
-// Licenced under the Apache Licence, Version 2.0 <LICENCE-APACHE or
-// http://www.apache.org/licenses/LICENSE-2.0> or the MIT license
-// <LICENCE-MIT or http://opensource.org/licenses/MIT>, at your
-// discretion. This file may not be copied, modified, or distributed
-// except according to those terms.
-
-pub mod dggrid;
