--- dggrs/src/bindings/wasm/mod.rs
@@ -1,4 +1,0 @@
-pub mod adapters;
-pub mod factory;
-pub mod models;
-pub mod ports;
